alx_travel_app/alx_travel_app/requirement.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=3.14
drf-yasg==1.21.10
celery>=5.3.1
django-environ>=0.9.0
mysqlclient>=2.1.0
